requirements.txt
flask
PyMuPDF
python_dotenv
google-generativeai==0.3.2
rank_bm25
tf-keras
sentence_transformers
biopython
faiss-cpu
rank_bm25